requirements.txt
google-genai
python-telegram-bot
python-dotenv